requirements_test.txt
pytest-homeassistant-custom-component==0.13.87
pre-commit